create a new Visibility interface to capture state visibility functionality
