+++ module/parser/src/commonMain/kotlin/org/cru/godtools/tool/model/Content.kt
@@ -1,7 +1,6 @@
 package org.cru.godtools.tool.model
 
 import org.cru.godtools.expressions.Expression
-import org.cru.godtools.expressions.toExpressionOrNull
 import org.cru.godtools.tool.ParserConfig
 import org.cru.godtools.tool.REGEX_SEQUENCE_SEPARATOR
 import org.cru.godtools.tool.internal.RestrictTo
@@ -10,22 +9,19 @@ import org.cru.godtools.tool.model.DeviceType.Companion.toDeviceTypes
 import org.cru.godtools.tool.model.tips.InlineTip
 import org.cru.godtools.tool.model.tips.Tip
 import org.cru.godtools.tool.model.tips.XMLNS_TRAINING
-import org.cru.godtools.tool.state.State
 import org.cru.godtools.tool.xml.XmlPullParser
 
 private const val XML_REQUIRED_FEATURES = "required-features"
 private const val XML_RESTRICT_TO = "restrictTo"
 private const val XML_VERSION = "version"
-private const val XML_INVISIBLE_IF = "invisible-if"
-private const val XML_GONE_IF = "gone-if"
 
-abstract class Content : BaseModel {
+abstract class Content : BaseModel, Visibility {
     private val version: Int
     private val restrictTo: Set<DeviceType>
     private val requiredFeatures: Set<String>
 
-    private val invisibleIf: Expression?
-    private val goneIf: Expression?
+    final override val invisibleIf: Expression?
+    final override val goneIf: Expression?
 
     internal constructor(parent: Base, parser: XmlPullParser) : super(parent) {
         version = parser.getAttributeValue(null, XML_VERSION)?.toIntOrNull() ?: SCHEMA_VERSION
@@ -33,8 +29,10 @@ abstract class Content : BaseModel {
         requiredFeatures = parser.getAttributeValue(XML_REQUIRED_FEATURES)
             ?.split(REGEX_SEQUENCE_SEPARATOR)?.filterTo(mutableSetOf()) { it.isNotBlank() }.orEmpty()
 
-        invisibleIf = parser.getAttributeValue(XML_INVISIBLE_IF).toExpressionOrNull()
-        goneIf = parser.getAttributeValue(XML_GONE_IF).toExpressionOrNull()
+        parser.parseVisibilityAttrs { invisibleIf, goneIf ->
+            this.invisibleIf = invisibleIf
+            this.goneIf = goneIf
+        }
     }
 
     @RestrictTo(RestrictToScope.TESTS)
@@ -62,11 +60,6 @@ abstract class Content : BaseModel {
             restrictTo.none { it in DeviceType.SUPPORTED } ||
             invisibleIf?.isValid() == false ||
             goneIf?.isValid() == false
-
-    fun isInvisible(state: State) = invisibleIf?.evaluate(state) ?: false
-    fun isInvisibleFlow(state: State) = state.changeFlow(invisibleIf?.vars()) { isInvisible(it) }
-    fun isGone(state: State) = goneIf?.evaluate(state) ?: false
-    fun isGoneFlow(state: State) = state.changeFlow(goneIf?.vars()) { isGone(it) }
 
     open val tips get() = emptyList<Tip>()
