+++ module/parser/src/iosMain/kotlin/org/cru/godtools/tool/model/IosVisibility.kt
@@ -4,9 +4,9 @@ import kotlinx.coroutines.flow.combine
 import org.cru.godtools.tool.state.State
 import org.cru.godtools.tool.util.watch
 
-fun Content.watchIsGone(state: State, block: (Boolean) -> Unit) = isGoneFlow(state).watch(block)
-fun Content.watchIsInvisible(state: State, block: (Boolean) -> Unit) = isInvisibleFlow(state).watch(block)
+fun Visibility.watchIsGone(state: State, block: (Boolean) -> Unit) = isGoneFlow(state).watch(block)
+fun Visibility.watchIsInvisible(state: State, block: (Boolean) -> Unit) = isInvisibleFlow(state).watch(block)
 
-fun Content.watchVisibility(state: State, block: (Boolean, Boolean) -> Unit) =
+fun Visibility.watchVisibility(state: State, block: (Boolean, Boolean) -> Unit) =
     isInvisibleFlow(state).combine(isGoneFlow(state)) { invisible, gone -> Pair(invisible, gone) }
         .watch { block(it.first, it.second) }
